pyproject.toml
[project]
name = "genai-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.37.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-groq>=1.1.2",
    "mysql-connector-python>=9.7.0",
    "plotly>=6.7.0",
    "pydantic>=2.13.3",
    "python-box>=7.4.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]
